Input: add SW_PEN_INSERTED define
Some devices with a pen may have a switch that can be used to detect
when the pen is inserted or removed to a slot on the device.  Let's add
a define to the input event codes so that everyone can be on the same
page for what event we should generate when the pen is inserted or
removed.

In general the pen switch could be used by the software on the device to
kick off any number of actions when the pen is inserted or removed.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
[ K9100ii - Value conflicts with Samsung custom event SW_GLOVE, which
  has usage within two (unused) touchscreen drivers, so give it a
  different dummy value ]
Signed-off-by: K9100ii <k9100ii.ki@gmail.com>