FIX&REVERTME: HID: playstation: disable CRC checking for calibration info when using Bluetooth
Fails on gta4xlwifi for some unknown reason, also with hid-sony, while
it doesn't on other devices. CRC checking with USB seems to work just
fine, so I guess it's something Bluetooth-related going wrong somewhere...

Unsure if DualSense is affected, but also disable the checking for
DualSense controllers anyway.

Apparently, the DS4 is meant to work as-is on the stock firmware on
these devices, according to videos on the stock firmware....

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>