Revert "BACKPORT: FROMGIT: kbuild: mkcompile_h: Include $LD version in /proc/version"
This reverts commit 2d578d836a8c886871add1943d6f824b2365e497.

When using clang with AMD's ROCm or NVIDIA's CUDA installed, when
reporting its version information (`clang -v`), it will report
information about installations of either. That commit causes that
information to be on a new line in setting the LINUX_COMPILER macro in
include/generated/compile.h, thereby breaking up the define and causing
compilation errors. For example:

 #define LINUX_COMPILER "Android (11209041, +pgo, +bolt, +lto, +mlgo, based on r510928) clang version 18.0.0 (https://android.googlesource.com/toolchain/llvm-project 477610d4d0d988e69dbc3fae4fe86bff3f07f2b5)
 Found HIP installation: /opt/rocm, version 5.7.31921, LLD 18.0.0"

A fair few more patches are necessary to make this commit properly
usable, but it doesn't seem worthwhile to backport and pull them in.

Linker information is also very far from critical - or typically
interesting - information, and in the case of Android kernels, the
linker being used is predictable from common usage - LLD with the
version corresponding to the clang version used.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>