usb: gadget: ncm: Revert "Handle decoding of multiple NTB's in unwrap call"
This reverts a port of commit 5bdf93a2f5459f944b416b188178ca4a92fd206f
alongside Samsung's NCM gadget driver changes, which doesn't work
properly and doesn't satisfy the original intentions of that commit.

Finally, I've tested USB tethering with NCM to find that out - For
reference, on the Android side, enabling it was setting two overlays
in a TetheringConfigOverlay RRO overlay:
 - "config_tether_usb_functions" set to 1 [1]
 - "config_tether_ncm_regexs" set with a `ncm\\d` item [2]
The RNDIS gadget has always been used for USB tethering, which remains
problem-free, but it's worth testing NCM nevertheless.

With a Linux PC, the reverted attempt of a port causes the "Wrong NDP
SIGN" condition to always be triggered. That can be got around by
switching the casting in tmp below the "walk through NDP" comment from
__le16 to void, also as per an otherwise same port attempt at [3], but
that doesn't look right at all, and more importantly, it's not so
different from not having the changes in that commit in any form at all.

 1: https://cs.android.com/android/platform/superproject/+/main:packages/modules/Connectivity/Tethering/res/values/config.xml;drc=fb65dfeb8da0c0ce4ec8ac63099ede54195e4da2;l=205
 2: https://cs.android.com/android/platform/superproject/+/main:packages/modules/Connectivity/Tethering/res/values/config.xml;drc=fb65dfeb8da0c0ce4ec8ac63099ede54195e4da2;l=34
 3: https://github.com/LineageOS/android_kernel_samsung_gta4xl/blob/2aef692b018bd42bd53c6fb7cbd1d2aa0dc2811d/drivers/usb/gadget/function/f_ncm.c#L1260

Most importantly though, that commit was intended to resolve issues with
Windows PCs. So, what happens with a Windows 11 PC?
With the port attempt, and with and without that casting change, the
"Wrong NDP SIGN" condition will always be triggered, meaning the issues
it was intended to resolve predictably haven't been.

I have no idea how to properly port it alongside Samsung's changes so it
is working properly, so simply revert it at least for now.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>