exynos-ufc: massively improve DVFS voltage control interface implementation
Dug deep within the git history is commit 30472c13ee ("exynos-ufc: add
CPU and general DVFS voltage control interfaces"). It was quite dirty,
with even a full unnecessary copy of a function (fvmap_copy_from_sram())
with minor changes, and I was meaning to clean it up since forever. This
finally does so.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>